Split Realtime.ts into multiple modules
This file was growing beyond what its original name implied, so let's
split it up.
